Refactor `EditExtraFragment`, `ExtraItem` and `Extras` to `ryey.easer.skills.reusable`.
