+++ app/src/main/java/ryey/easer/skills/reusable/ExtraItem.java
@@ -17,7 +17,7 @@
  * along with Easer.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ryey.easer.skills.operation;
+package ryey.easer.skills.reusable;
 
 import android.os.Parcel;
 import android.os.Parcelable;
