+++ app/src/main/res/layout/skill_reusable__intent_fragment_extra_item.xml
